// src/app/model/diagnosticoO.ts
/**
 * Proyecto TDS Backend
 * Descripción
 *
 * OpenAPI spec version: 2.0
 * Contact: info@tecazuay.edu.ec
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Odontologia } from './odontologia';

export interface DiagnosticoO { 
    cie1?: string;
    cie2?: string;
    def?: string;
    def2?: string;
    descripcion1?: string;
    descripcion2?: string;
    fechaA?: Date;
    fechaC?: Date;
    idDiagnostico?: number;
    odontologia?: Odontologia;
    pre1?: string;
    pre2?: string;
    profesional?: string;
}