// src/app/model/examenEsto.ts
/**
 * Proyecto TDS Backend
 * Descripción
 *
 * OpenAPI spec version: 2.0
 * Contact: info@tecazuay.edu.ec
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Odontologia } from './odontologia';

export interface ExamenEsto { 
    atm?: boolean;
    carrillos?: boolean;
    descripcion?: string;
    glandulasSa?: boolean;
    glangios?: boolean;
    idExamenEs?: number;
    labio?: boolean;
    lengua?: boolean;
    maxSuperior?: boolean;
    maxilarInf?: boolean;
    mejillas?: boolean;
    odontologia?: Odontologia;
    oroFaringe?: boolean;
    paladar?: boolean;
    piso?: boolean;
}