// src/app/model/indicesCPO.ts
/**
 * Proyecto TDS Backend
 * Descripción
 *
 * OpenAPI spec version: 2.0
 * Contact: info@tecazuay.edu.ec
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Odontologia } from './odontologia';

export interface IndicesCPO { 
    c?: number;
    c1?: number;
    d?: number;
    d1?: number;
    idIndice?: number;
    o?: number;
    o1?: number;
    odontologia?: Odontologia;
    p?: number;
    p1?: number;
    totalD?: number;
    totald1?: number;
}